"Skrev" om ftrl/saksbehandling.js til .tsx
